+++ CHANGELOG.md
@@ -9,6 +9,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Fixed
 
 - Fixed invalid variable names created for instance-based auto-imports
+- Fixed diagnostics not getting updated for dependent files when a file is edited and workspace diagnostics is
+  enabled ([#1042](https://github.com/JohnnyMorganz/luau-lsp/issues/1042))
 
 ## [1.45.0] - 2025-05-04